Agrego boton eliminar y protego rutas

+++ api_clinica/medicos.js
@@ -2,9 +2,10 @@ import express from "express";
 import { db } from "./db.js";
 import { body, query } from "express-validator";
 import { validarId, verificarValidaciones, validarMedico } from "./validaciones.js";
+import passport from "passport";
 const router = express.Router();
 
-
+router.use(passport.authenticate("jwt", { session: false }));
 
 // ------------GET------------
 router.get('/', async (req, res) => {
@@ -65,6 +66,11 @@ router.put(
         const id = Number(req.params.id);
         const { nombre, apellido, especialidad, matricula } = req.body;
 
+        const [rows] = await db.execute("SELECT * FROM medicos WHERE id = ?", [id]);
+        if (rows.length === 0) {
+            return res.status(404).json({ success: false, message: "Médico no encontrado" });
+        }
+
         await db.execute(
             "UPDATE medicos SET nombre=?, apellido=?, especialidad=?, matricula=?  WHERE id=?",
             [nombre, apellido, especialidad, matricula, id]
@@ -85,6 +91,11 @@ router.put(
 
 router.delete("/:id", validarId, verificarValidaciones, async (req, res) => {
     const id = Number(req.params.id);
+
+    const [rows] = await db.execute("SELECT * FROM medicos WHERE id = ?", [id]);
+    if (rows.length === 0) {
+        return res.status(404).json({ success: false, message: "Médico no encontrado" });
+    }
 
     await db.execute("DELETE FROM medicos WHERE id=?", [id]);
     res.json({ success: true, data: id });
