+++ api_clinica/pacientes.js
@@ -2,8 +2,10 @@ import express from "express";
 import { db } from "./db.js";
 import { body, query } from "express-validator";
 import { validarId, verificarValidaciones, validarPacientes } from "./validaciones.js";
+import passport from "passport";
 const router = express.Router();
 
+router.use(passport.authenticate("jwt", { session: false }));
 
 // ------------GET------------
 router.get('/', async (req, res) => {
@@ -20,12 +22,13 @@ router.get('/', async (req, res) => {
 router.get('/:id', validarId, verificarValidaciones, async (req, res) => {
     const { id } = req.params;
 
+
     const [rows] = await db.execute("SELECT * FROM pacientes WHERE id = ?", [id]);
 
     if (rows.length === 0) {
         return res
             .status(404)
-            .json({ success: false, message: "Paciente no encontrado" });
+            .json({ success: false, error: "Paciente no encontrado" });
     }
 
     res.json({ success: true, data: rows[0] });
@@ -64,6 +67,11 @@ router.put(
         const id = Number(req.params.id);
         const { nombre, apellido, DNI, nacimiento, obraSocial } = req.body;
 
+        const [existe] = await db.execute("SELECT * FROM usuarios WHERE id=?", [id]);
+        if (existe.length === 0) {
+            return res.status(404).json({ success: false, error: "Usuario no encontrado" });
+        }
+
         await db.execute(
             "UPDATE pacientes SET nombre=?, apellido=?, DNI=?, nacimiento=?, obraSocial=?  WHERE id=?",
             [nombre, apellido, DNI, nacimiento, obraSocial, id]
@@ -84,6 +92,11 @@ router.put(
 
 router.delete("/:id", validarId, verificarValidaciones, async (req, res) => {
     const id = Number(req.params.id);
+
+    const [existe] = await db.execute("SELECT * FROM usuarios WHERE id=?", [id]);
+    if (existe.length === 0) {
+        return res.status(404).json({ success: false, error: "Usuario no encontrado" });
+    }
 
     await db.execute("DELETE FROM pacientes WHERE id=?", [id]);
     res.json({ success: true, data: id });
